(chore) move LogDetail helper files into lib subfolder
