+++ client/src/components/logbooks/LogDetail/lib/build-new-field-values.ts
