+++ client/src/components/logbooks/LogDetail/lib/get-fields.ts
