+++ client/src/components/logbooks/LogDetail/lib/get-rows.ts
