+++ client/src/components/logbooks/LogDetail/lib/has-values.ts
