+++ client/src/components/logbooks/LogDetail/lib/is-valid.ts
